wip: keep SuperCollider stuff in a separate branch
I had a lot of fun configuring scsynth.wasm, but it seems like there's not a single good client package for compiling ugens in JavaScript, and it's a risk relying on an old ScalaCollider package.

I have to make a working app for Uni in time, so I think it's ok to put back this idea for now. Might bring it back at some point