refactor(ui-kit): remove `Object.assign` for compound components
There's a problem with compound components in Next.js. When a client
component gets serialized for SSR (or something), it gets temporarily
replaced with a 'lazy React component', so my `__slotName__` gets lost

I've chosen a path where the `.Root` component is needed, because it's
somewhat reasonable to indicate that the component is compound. Then
it's possible to use regular `export * as` clause that the bundler can
reason about

The issue might be half fixed yet because of `__slotName__`